move app stuff around more
