+++ examples/platform/nrf528xx/app/Server.cpp
@@ -50,7 +50,7 @@
 #include "attribute-storage.h"
 #include "gen/znet-bookkeeping.h"
 #include "util.h"
-#include "util/chip-zcl-zpro-codec.h"
+#include "app/chip-zcl-zpro-codec.h"
 
 using namespace ::chip;
 using namespace ::chip::Inet;
