+++ src/app/DataModel.am
@@ -24,7 +24,7 @@
 #      must be anchored relative to the top build directory.
 #
 
-CHIP_BUILD_DATA_MODEL_SOURCE_FILES                         = \
-    @top_builddir@/src/app/util/encoder.c                    \
-    @top_builddir@/src/app/util/decoder.c                    \
+CHIP_BUILD_DATA_MODEL_SOURCE_FILES                    = \
+    @top_builddir@/src/app/encoder.c                    \
+    @top_builddir@/src/app/decoder.c                    \
     $(NULL)
