+++ src/app/encoder.c
